Add admin checks to server functions

+++ src/actions/module.mjs
@@ -3,10 +3,15 @@
 import ModuleDao from '../dao/module.mjs';
 import Controller from '../controllers/controller.mjs';
 import db from '../db/db.mjs';
+import useLoggedIn from '../hooks/login.mjs';
 import xss from 'xss';
 import fs from 'node:fs/promises';
 
 export async function addModule(prevState, formData) {
+    const { isAdmin } = await useLoggedIn();
+    if(!isAdmin) {
+        return {"error" : "Only admins can add a module."};
+    }
     const code = formData.get("moduleCode"), name = formData.get("moduleName");
     if(code && name && code.match("^\\w+$")) {
         const moduleDao = new ModuleDao(db);
@@ -32,6 +37,10 @@ export async function addModule(prevState, formData) {
 }
 
 export async function deleteModule(id) {
+    const { isAdmin } = await useLoggedIn();
+    if(!isAdmin) {
+        return {"error" : "Only admins can delete a module."};
+    }
     const controller = new Controller(db);
     const module = controller.moduleDao.getModuleById(id);
     const result = controller.deleteModule(id);
